+++ src/main/kotlin/com/marketboro/usecase/dto/TotalPointsDto.kt
@@ -1,5 +1,5 @@
 package com.marketboro.usecase.dto
 
 data class TotalPointsDto(
-    val totalPoints: Long
+    val totalPoints: Int
 )
